extract VmInterrupt as well
