+++ profiling/src/profiling/interrupt_manager.rs
@@ -1,6 +1,25 @@
-use super::VmInterrupt;
-use std::sync::atomic::Ordering;
+use std::sync::atomic::{AtomicBool, AtomicU32, Ordering};
 use std::sync::Mutex;
+
+#[derive(Debug, Eq, PartialEq, Hash)]
+pub struct VmInterrupt {
+    pub interrupt_count_ptr: *const AtomicU32,
+    pub engine_ptr: *const AtomicBool,
+}
+
+impl std::fmt::Display for VmInterrupt {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "VmInterrupt{{{:?}, {:?}}}",
+            self.interrupt_count_ptr, self.engine_ptr
+        )
+    }
+}
+
+// This is a lie, technically, but we're trying to build it safely on top of
+// the PHP VM.
+unsafe impl Send for VmInterrupt {}
 
 pub(super) struct InterruptManager {
     vm_interrupts: Mutex<Vec<VmInterrupt>>,
